requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
imaplib2==2.45.0
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
openai==1.3.0
spacy==3.7.2
nltk==3.8.1
schedule==1.2.0
